% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{clean_path}
\alias{clean_path}
\title{Clean path by removing leading slashes}
\usage{
clean_path(path)
}
\arguments{
\item{path}{File path}
}
\value{
Cleaned path
}
\description{
Clean path by removing leading slashes
}
\keyword{internal}
