% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{connect_to_sharepoint}
\alias{connect_to_sharepoint}
\title{Connect to SharePoint site}
\usage{
connect_to_sharepoint(sp_conf)
}
\arguments{
\item{sp_conf}{SharePoint configuration list}
}
\value{
List with token and drive_id
}
\description{
This function orchestrates the connection process:
\enumerate{
\item Gets an access token
\item Retrieves the SharePoint site information
\item Gets the document library (drive) ID
}
}
\keyword{internal}
