% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{download_file_from_sharepoint}
\alias{download_file_from_sharepoint}
\title{Download file from SharePoint drive}
\usage{
download_file_from_sharepoint(remote_path, dest, drive_id, token)
}
\arguments{
\item{remote_path}{Path to file in SharePoint}

\item{dest}{Local destination path}

\item{drive_id}{SharePoint drive ID}

\item{token}{Microsoft Graph API access token}
}
\description{
Downloads a file from SharePoint document library to local filesystem.
}
\keyword{internal}
