% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo-ingestion.R
\name{flatten_field}
\alias{flatten_field}
\title{Flatten Survey Data Fields}
\usage{
flatten_field(x, p)
}
\arguments{
\item{x}{A vector or list representing a field in the data.}

\item{p}{The prefix or name associated with the field, used for naming during the flattening process.}
}
\value{
Modified field, either unchanged, unnested, or appropriately renamed.
}
\description{
Processes each field within a row of survey data, handling both simple vectors and nested lists. For lists with named elements, renames and unlists them for flat structure preparation.
}
\keyword{internal}
