% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo-ingestion.R
\name{flatten_row}
\alias{flatten_row}
\title{Flatten Survey Data Rows}
\usage{
flatten_row(x)
}
\arguments{
\item{x}{A list representing a row of data, potentially containing nested lists or vectors.}
}
\value{
A tibble with each row representing flattened survey data.
}
\description{
Transforms each row of nested survey data into a flat tabular format using a mapping and flattening process.
}
\keyword{internal}
