% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_sharepoint_site}
\alias{get_sharepoint_site}
\title{Get SharePoint site information}
\usage{
get_sharepoint_site(site_url, token)
}
\arguments{
\item{site_url}{Full SharePoint site URL}

\item{token}{Microsoft Graph API access token}
}
\value{
Site information as a list
}
\description{
Retrieves site metadata from Microsoft Graph API using the site URL.
}
\keyword{internal}
