% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_site_drive_id}
\alias{get_site_drive_id}
\title{Get SharePoint site drive ID}
\usage{
get_site_drive_id(site_id, token)
}
\arguments{
\item{site_id}{SharePoint site ID}

\item{token}{Microsoft Graph API access token}
}
\value{
Drive ID string
}
\description{
Retrieves the default document library (drive) ID for a SharePoint site.
}
\keyword{internal}
