% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_dotenv}
\alias{load_dotenv}
\title{Load environment variables from .env file}
\usage{
load_dotenv(file = ".env")
}
\arguments{
\item{file}{Path to .env file, defaults to ".env" in current working directory}
}
\value{
NULL (called for side effects)
}
\description{
Loads environment variables from .env file if it exists. This function should
be called before reading configuration to ensure dotenv variables are available.
}
\keyword{helper}
