% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZooGoN-package.R
\docType{package}
\name{ZooGoN-package}
\alias{ZooGoN}
\alias{ZooGoN-package}
\title{ZooGoN: Gulf of Naples 40-Year Zooplankton Biodiversity Assessment and Standardization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Standardizes taxonomic names in zooplankton datasets from the LTER-MareChiara station in the Gulf of Naples (1984-2024). This package is part of the DTO-BioFlow project (Digital Twin Ocean - Biodiversity Flow Integration) under the EU Horizon Mission to restore ocean and waters by 2030. It processes 40 years of Mediterranean zooplankton monitoring data, extracting standardized genus-species information and converting datasets to Darwin Core Archive format for integration with EMODnet Biology and the European Digital Twin of the Ocean.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ioledc/ZOOGoN-40Y}
  \item \url{https://ioledc.github.io/ZOOGoN-40Y/}
  \item Report bugs at \url{https://github.com/ioledc/ZOOGoN-40Y/issues}
}

}
\author{
\strong{Maintainer}: Lorenzo Longobardi \email{lorenzo.longobardi@gmail.com} (\href{https://orcid.org/0000-0003-3126-7341}{ORCID})

Authors:
\itemize{
  \item Iole Di Capua \email{iole.dicapua@szn.it} (\href{https://orcid.org/0000-0003-2959-8977}{ORCID})
}

}
\keyword{internal}
