% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo-ingestion.R
\name{get_kobo_data}
\alias{get_kobo_data}
\title{Retrieve Data from Kobotoolbox API}
\usage{
get_kobo_data(
  assetid,
  url = "eu.kobotoolbox.org",
  uname = NULL,
  pwd = NULL,
  encoding = "UTF-8",
  format = "json"
)
}
\arguments{
\item{assetid}{The asset ID of the Kobotoolbox form.}

\item{url}{The URL of Kobotoolbox (default is "eu.kobotoolbox.org").}

\item{uname}{Username for Kobotoolbox account.}

\item{pwd}{Password for Kobotoolbox account.}

\item{encoding}{Encoding to be used for data retrieval (default is "UTF-8").}

\item{format}{Format of the data to retrieve, either "json" or "xml" (default is "json").}
}
\value{
A list containing all retrieved survey results.
}
\description{
This function retrieves survey data from Kobotoolbox API for a specific asset.
It supports pagination and handles both JSON and XML formats.
}
\details{
The function uses pagination to retrieve large datasets, with a limit of 30,000 records per request.
It continues to fetch data until all records are retrieved or an error occurs.
}
\examples{
\dontrun{
kobo_data <- get_kobo_data(
  assetid = "your_asset_id",
  uname = "your_username",
  pwd = "your_password"
)
}
}
\keyword{ingestion}
