% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-surveys.R
\name{preprocess_surveys}
\alias{preprocess_surveys}
\title{Preprocess MC Surveys}
\usage{
preprocess_surveys(raw_data = NULL)
}
\arguments{
\item{raw_data}{Data to preprocess.}
}
\value{
Preprocessed data
}
\description{
This function preprocesses raw MC survey data.
It performs data cleaning and transformation
}
\examples{
\dontrun{
preprocess_surveys()
}
}
\keyword{preprocessing}
\keyword{workflow}
