% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion-surveys.R
\name{rename_child}
\alias{rename_child}
\title{Rename Nested Survey Data Elements}
\usage{
rename_child(x, i, p)
}
\arguments{
\item{x}{A list element, possibly nested, to be renamed.}

\item{i}{The index or key of the element within the parent list.}

\item{p}{The parent name to prepend to the element's existing name for context.}
}
\value{
A renamed list element, structured to maintain contextual relevance in a flattened dataset.
}
\description{
Appends a parent name or index to child elements within a nested list, assisting in creating a coherent and traceable data structure during the flattening process.
}
\keyword{internal}
