% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{upload_sharepoint_df}
\alias{upload_sharepoint_df}
\title{Upload a data frame to SharePoint}
\usage{
upload_sharepoint_df(
  data,
  prefix,
  options,
  format = c("csv", "tsv", "parquet")
)
}
\arguments{
\item{data}{A data frame to upload}

\item{prefix}{File prefix path (e.g., "raw", "preprocessed")}

\item{options}{SharePoint configuration list from config$storage$sharepoint}

\item{format}{File format: "csv", "tsv", or "parquet". Default is "csv"}
}
\value{
Invisible NULL
}
\description{
Upload a data frame to SharePoint
}
\examples{
\dontrun{
conf <- read_config()
upload_sharepoint_df(my_data, "raw", conf$storage$sharepoint)
upload_sharepoint_df(my_data, "preprocessed", conf$storage$sharepoint, format = "parquet")
}

}
\keyword{storage}
